third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/access-control-basic-allow-access-control-origin-header.htm

<!DOCTYPE html>
<html>
  <head>
    <title></title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script type="text/javascript">
    async_test(function(test) {
      const xhr = new XMLHttpRequest;

      try {
        xhr.open("GET", get_host_info().HTTP_REMOTE_ORIGIN + "/XMLHttpRequest/resources/access-control-origin-header.py", false);
      } catch(e) {
        assert_unreached("Exception thrown. Cross-domain access is not allowed in 'open'. [" + e.message + "].");
        return;
      }

      try {
        xhr.send();
      } catch(e) {
        assert_unreached("Exception thrown. Cross-domain access is not allowed in 'send'. [" + e.message + "].");
        return;
      }

      assert_equals(xhr.responseText.replace("www1.", ""), "PASS: Cross-domain access allowed.\n" +
          "HTTP_ORIGIN: " + get_host_info().HTTP_REMOTE_ORIGIN.replace("www1.", ""));
      test.done();
    }, "Access control test with origin header");
    </script>
  </body>
</html>
